Revert "Move function_schema to ATen/core (#12467)"
This reverts commit caf8b0777a63a694393da1213f21c0a6a08c00b6.
